HTML and CSS code
The standard calculator is coded in HTML and CSS. Script tags are added to perform basic addition, subtraction, multiplication, and division. A clear button is added to clear the screen completely, and a delete button is added to remove a single number at a time. If the function is incorrect and we enter an equal value, it will indicate that the result is invalid.